requirements.txt
inflect==6.0.0
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.1
django-filter==2.4.0
flake8==5.0.4
isort==5.10.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
